修复大屏导出失败问题

See merge request simo/simo-leaderview!777

+++ simo-leaderview/src/main/java/com/uxsino/leaderview/dao/IHomeTemplateImgDao.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 public interface IHomeTemplateImgDao extends ICustomRepository<HomeTemplateImg, Long> {
 
     @Query("select max(id) from HomeTemplateImg ")
-    Integer getMaxId();
+    Long getMaxId();
 
 }
